implement test_scalar_compat, move shift test
